requirements.txt
beautifulsoup4==4.9.3
google-api-python-client==2.108.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
oauthlib==3.1.0
requests-oauthlib==1.3.0
redis==3.5.3
flask==3.0.0
PyYAML>=5.4.1
oauth2client==4.1.3
lark-oapi==1.0.26


# 分类
jieba==0.42.1
joblib==1.3.2 
scikit-learn==1.3.2 
scipy==1.11.4 
threadpoolctl==3.2.0